+++ core/src/main/java/org/apache/metamodel/query/ScalarFunction.java
@@ -33,11 +33,13 @@ public interface ScalarFunction extends FunctionType {
      * 
      * @param row
      *            the row containing data
+     * @param parameters
+     *            any parameters associated with the function call
      * @param operandItem
      *            the select item which is the argument to this function. If a
      *            function takes multiple select items, this will be the primary
      *            one.
      * @return
      */
-    public Object evaluate(Row row, SelectItem operandItem);
+    public Object evaluate(Row row, Object[] parameters, SelectItem operandItem);
 }