+++ core/src/main/java/org/apache/metamodel/query/ToBooleanFunction.java
@@ -38,7 +38,7 @@ public class ToBooleanFunction extends DefaultScalarFunction {
     }
 
     @Override
-    public Object evaluate(Row row, SelectItem item) {
+    public Object evaluate(Row row, Object[] parameters, SelectItem item) {
         final Object value = row.getValue(item);
         if (value == null || value instanceof Boolean) {
             return value;