+++ core/src/test/java/org/apache/metamodel/query/parser/QueryParserTest.java
@@ -68,6 +68,18 @@ public class QueryParserTest extends TestCase {
         assertEquals("SELECT TO_NUMBER(a.foo) FROM sch.tbl a WHERE TO_BOOLEAN(a.bar) = FALSE", q.toSql());
     }
 
+    public void testSelectMapValueUsingDotNotation() throws Exception {
+        // set 'baz' column to a MAP column
+        MutableColumn col = (MutableColumn) dc.getColumnByQualifiedLabel("tbl.baz");
+        col.setType(ColumnType.MAP);
+
+        Query q = MetaModelHelper.parseQuery(dc,
+                "SELECT sch.tbl.baz.foo.bar, baz.helloworld, baz.hello.world FROM sch.tbl");
+        assertEquals(
+                "SELECT MAP_VALUE('foo.bar',tbl.baz), MAP_VALUE('helloworld',tbl.baz), MAP_VALUE('hello.world',tbl.baz) FROM sch.tbl",
+                q.toSql());
+    }
+
     public void testSelectEverythingFromTable() throws Exception {
         Query q = MetaModelHelper.parseQuery(dc, "SELECT * FROM sch.tbl");
         assertEquals("SELECT tbl.foo, tbl.bar, tbl.baz FROM sch.tbl", q.toSql());
@@ -221,7 +233,8 @@ public class QueryParserTest extends TestCase {
 
         q = MetaModelHelper.parseQuery(dc,
                 "SELECT COUNT(*) FROM sch.tbl a LEFT JOIN sch.tbl b ON a.foo = b.foo AND a.bar = b.baz");
-        assertEquals("SELECT COUNT(*) FROM sch.tbl a LEFT JOIN sch.tbl b ON a.foo = b.foo AND a.bar = b.baz", q.toSql());
+        assertEquals("SELECT COUNT(*) FROM sch.tbl a LEFT JOIN sch.tbl b ON a.foo = b.foo AND a.bar = b.baz",
+                q.toSql());
     }
 
     public void testSimpleSelectFromWhere() throws Exception {
@@ -230,11 +243,11 @@ public class QueryParserTest extends TestCase {
 
         FilterClause whereClause = q.getWhereClause();
         assertEquals(2, whereClause.getItemCount());
-        assertNull("WHERE item was an expression based item, which indicates it was not parsed", whereClause.getItem(0)
-                .getExpression());
+        assertNull("WHERE item was an expression based item, which indicates it was not parsed",
+                whereClause.getItem(0).getExpression());
         assertEquals(2, whereClause.getItemCount());
-        assertNull("WHERE item was an expression based item, which indicates it was not parsed", whereClause.getItem(1)
-                .getExpression());
+        assertNull("WHERE item was an expression based item, which indicates it was not parsed",
+                whereClause.getItem(1).getExpression());
 
         assertEquals("baz", whereClause.getItem(0).getOperand());
         assertEquals(Integer.class, whereClause.getItem(1).getOperand().getClass());
@@ -264,8 +277,8 @@ public class QueryParserTest extends TestCase {
     }
 
     public void testCoumpoundWhereClause() throws Exception {
-        Query q = MetaModelHelper
-                .parseQuery(dc, "SELECT foo FROM sch.tbl WHERE (bar = 'baz' OR (baz > 5 AND baz < 7))");
+        Query q = MetaModelHelper.parseQuery(dc,
+                "SELECT foo FROM sch.tbl WHERE (bar = 'baz' OR (baz > 5 AND baz < 7))");
         assertEquals("SELECT tbl.foo FROM sch.tbl WHERE (tbl.bar = 'baz' OR (tbl.baz > 5 AND tbl.baz < 7))", q.toSql());
 
         FilterClause wc = q.getWhereClause();
@@ -281,8 +294,8 @@ public class QueryParserTest extends TestCase {
     }
 
     public void testCoumpoundWhereClauseDelimInLoweCase() throws Exception {
-        Query q = MetaModelHelper
-                .parseQuery(dc, "SELECT foo FROM sch.tbl WHERE (bar = 'baz' OR (baz > 5 and baz < 7))");
+        Query q = MetaModelHelper.parseQuery(dc,
+                "SELECT foo FROM sch.tbl WHERE (bar = 'baz' OR (baz > 5 and baz < 7))");
         assertEquals("SELECT tbl.foo FROM sch.tbl WHERE (tbl.bar = 'baz' OR (tbl.baz > 5 AND tbl.baz < 7))", q.toSql());
 
         FilterClause wc = q.getWhereClause();
@@ -302,8 +315,8 @@ public class QueryParserTest extends TestCase {
         assertEquals("SELECT tbl.foo FROM sch.tbl WHERE tbl.bar IS NULL", q.toSql());
 
         assertEquals(1, q.getWhereClause().getItemCount());
-        assertNull("WHERE item was an expression based item, which indicates it was not parsed", q.getWhereClause()
-                .getItem(0).getExpression());
+        assertNull("WHERE item was an expression based item, which indicates it was not parsed",
+                q.getWhereClause().getItem(0).getExpression());
         assertNull(q.getWhereClause().getItem(0).getOperand());
         assertEquals(OperatorType.EQUALS_TO, q.getWhereClause().getItem(0).getOperator());
     }
@@ -313,8 +326,8 @@ public class QueryParserTest extends TestCase {
         assertEquals("SELECT tbl.foo FROM sch.tbl WHERE tbl.bar IS NOT NULL", q.toSql());
 
         assertEquals(1, q.getWhereClause().getItemCount());
-        assertNull("WHERE item was an expression based item, which indicates it was not parsed", q.getWhereClause()
-                .getItem(0).getExpression());
+        assertNull("WHERE item was an expression based item, which indicates it was not parsed",
+                q.getWhereClause().getItem(0).getExpression());
         assertNull(q.getWhereClause().getItem(0).getOperand());
         assertEquals(OperatorType.DIFFERENT_FROM, q.getWhereClause().getItem(0).getOperator());
     }
@@ -407,7 +420,8 @@ public class QueryParserTest extends TestCase {
                         + "GROUP BY foo HAVING COUNT(*) > 2 ORDER BY foo LIMIT 20 OFFSET 10");
         assertEquals(
                 "SELECT tbl.foo, COUNT(*), MAX(tbl.baz) FROM sch.tbl WHERE tbl.bar = 'baz' AND tbl.foo = tbl.bar AND tbl.baz > 5 "
-                        + "GROUP BY tbl.foo HAVING COUNT(*) > 2 ORDER BY tbl.foo ASC", q.toSql());
+                        + "GROUP BY tbl.foo HAVING COUNT(*) > 2 ORDER BY tbl.foo ASC",
+                q.toSql());
         assertEquals(20, q.getMaxRows().intValue());
         assertEquals(11, q.getFirstRow().intValue());