+++ elasticsearch/rest/src/test/java/org/apache/metamodel/elasticsearch/rest/JestElasticSearchDataContextTest.java
@@ -36,6 +36,7 @@ import org.apache.metamodel.elasticsearch.rest.utils.EmbeddedElasticsearchServer
 import org.apache.metamodel.query.FunctionType;
 import org.apache.metamodel.query.Query;
 import org.apache.metamodel.query.SelectItem;
+import org.apache.metamodel.query.parser.QueryParserException;
 import org.apache.metamodel.schema.Column;
 import org.apache.metamodel.schema.ColumnType;
 import org.apache.metamodel.schema.Schema;
@@ -539,7 +540,7 @@ public class JestElasticSearchDataContextTest {
         dataContext.query().from("nonExistingTable").select("user").and("message").execute();
     }
 
-    @Test(expected = IllegalArgumentException.class)
+    @Test(expected = QueryParserException.class)
     public void testQueryForAnExistingTableAndNonExistingField() throws Exception {
         indexTweeterDocument(indexType1, 1);
         dataContext.query().from(indexType1).select("nonExistingField").execute();